:sparkles: Refactor workouts
