+++ lib/workoutDetail_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:fitness_coaching_application_test/pose_detector_view.dart';
+import './workouts/pose_detector_view.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:ionicons/ionicons.dart';
 import 'color.dart';
