+++ lib/workouts/pose_painter.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:google_ml_kit/google_ml_kit.dart';
 
-import 'coordinates_translator.dart';
+import '../coordinates_translator.dart';
 
 class PosePainter extends CustomPainter {
   // PosePainter(this.poses, this.absoluteImageSize, this.rotation, this.armAngle, this.elbowAngle);
