+++ lib/workouts/workoutLandscapeStepPauseCamera_view.dart
@@ -1,17 +1,49 @@
+import 'package:camera/camera.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:fitness_coaching_application_test/color.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:google_ml_kit/google_ml_kit.dart';
 import 'package:ionicons/ionicons.dart';
-import 'color.dart';
 
-class WorkoutLandscapeStepPause extends StatefulWidget {
-  const WorkoutLandscapeStepPause({Key? key}) : super(key: key);
+import '../main.dart';
+
+class WorkoutLandscapeStepPauseCamera extends StatefulWidget {
+  WorkoutLandscapeStepPauseCamera(
+      {Key? key,
+      required this.title,
+      required this.customPaint,
+      required this.onImage,
+      this.initialDirection = CameraLensDirection.back})
+      : super(key: key);
+
+  final String title;
+  final CustomPaint? customPaint;
+  final Function(InputImage inputImage) onImage;
+  final CameraLensDirection initialDirection;
+
   @override
-  _WorkoutLandscapeStepPauseState createState() =>
-      _WorkoutLandscapeStepPauseState();
+  _WorkoutLandscapeStepPauseCameraState createState() =>
+      _WorkoutLandscapeStepPauseCameraState();
 }
 
-class _WorkoutLandscapeStepPauseState extends State<WorkoutLandscapeStepPause> {
+class _WorkoutLandscapeStepPauseCameraState
+    extends State<WorkoutLandscapeStepPauseCamera> {
+  CameraController? _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _startLiveFeed();
+  }
+
+  @override
+  void dispose() {
+    _stopLiveFeed();
+    super.dispose();
+  }
+
   Widget _landscapeMode() {
     SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersiveSticky);
     return Column(
@@ -78,12 +110,17 @@ class _WorkoutLandscapeStepPauseState extends State<WorkoutLandscapeStepPause> {
           ],
         ),
         Stack(children: <Widget>[
-          Container(
-            decoration: new BoxDecoration(color: color_white),
-            height: MediaQuery.of(context).size.height -
-                (MediaQuery.of(context).size.height * 0.25),
-            width: MediaQuery.of(context).size.width,
-          ),
+          Stack(children: <Widget>[
+            Container(
+                decoration: new BoxDecoration(color: color_white),
+                height: MediaQuery.of(context).size.height -
+                    (MediaQuery.of(context).size.height * 0.25),
+                width: MediaQuery.of(context).size.width,
+                child: Stack(fit: StackFit.expand, children: <Widget>[
+                  CameraPreview(_controller!),
+                  if (widget.customPaint != null) widget.customPaint!,
+                ]))
+          ]),
           Positioned(
               child: Center(
                   child: Container(
@@ -192,16 +229,78 @@ class _WorkoutLandscapeStepPauseState extends State<WorkoutLandscapeStepPause> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: OrientationBuilder(
-        builder: (context, orientation) {
-          if (orientation == Orientation.portrait) {
-            return _portraitMode();
-          } else {
-            return _landscapeMode();
-          }
-        },
-      ),
+    return Scaffold(body: OrientationBuilder(
+      builder: (context, orientation) {
+        if (orientation == Orientation.portrait) {
+          return _portraitMode();
+        } else {
+          return _landscapeMode();
+        }
+      },
+    ));
+  }
+
+  Future _startLiveFeed() async {
+    final camera = cameras[1];
+    _controller = CameraController(
+      camera,
+      ResolutionPreset.low,
+      enableAudio: false,
     );
+    _controller?.initialize().then((_) {
+      if (!mounted) {
+        return;
+      }
+      _controller?.startImageStream(_processCameraImage);
+      setState(() {});
+    });
+  }
+
+  Future _stopLiveFeed() async {
+    await _controller?.stopImageStream();
+    await _controller?.dispose();
+    _controller = null;
+  }
+
+  Future _processCameraImage(CameraImage image) async {
+    final WriteBuffer allBytes = WriteBuffer();
+    for (Plane plane in image.planes) {
+      allBytes.putUint8List(plane.bytes);
+    }
+    final bytes = allBytes.done().buffer.asUint8List();
+
+    final Size imageSize =
+        Size(image.width.toDouble(), image.height.toDouble());
+
+    final camera = cameras[1];
+    final imageRotation =
+        InputImageRotationMethods.fromRawValue(camera.sensorOrientation) ??
+            InputImageRotation.Rotation_0deg;
+
+    final inputImageFormat =
+        InputImageFormatMethods.fromRawValue(image.format.raw) ??
+            InputImageFormat.NV21;
+
+    final planeData = image.planes.map(
+      (Plane plane) {
+        return InputImagePlaneMetadata(
+          bytesPerRow: plane.bytesPerRow,
+          height: plane.height,
+          width: plane.width,
+        );
+      },
+    ).toList();
+
+    final inputImageData = InputImageData(
+      size: imageSize,
+      imageRotation: imageRotation,
+      inputImageFormat: inputImageFormat,
+      planeData: planeData,
+    );
+
+    final inputImage =
+        InputImage.fromBytes(bytes: bytes, inputImageData: inputImageData);
+
+    widget.onImage(inputImage);
   }
 }
