+++ lib/workouts/workoutPortraitStepBeginCamera_view.dart
@@ -8,8 +8,8 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:google_ml_kit/google_ml_kit.dart';
 import 'package:image_picker/image_picker.dart';
 
-import 'exerciseSumFinished_view.dart';
-import 'main.dart';
+import '../exerciseSumFinished_view.dart';
+import '../main.dart';
 
 enum ScreenMode { liveFeed, gallery }
