+++ lib/workouts/workoutPortraitStepFinishCamera_view.dart
@@ -5,11 +5,12 @@ import 'package:flutter/services.dart';
 import 'package:fitness_coaching_application_test/color.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:google_ml_kit/google_ml_kit.dart';
+import 'package:ionicons/ionicons.dart';
 
-import 'main.dart';
+import '../main.dart';
 
-class WorkoutLandscapeStepCountingCamera extends StatefulWidget {
-  WorkoutLandscapeStepCountingCamera(
+class WorkoutPortraitStepFinishCamera extends StatefulWidget {
+  WorkoutPortraitStepFinishCamera(
       {Key? key,
       required this.title,
       required this.customPaint,
@@ -23,12 +24,12 @@ class WorkoutLandscapeStepCountingCamera extends StatefulWidget {
   final CameraLensDirection initialDirection;
 
   @override
-  _WorkoutLandscapeStepCountingCameraState createState() =>
-      _WorkoutLandscapeStepCountingCameraState();
+  _WorkoutPortraitStepFinishCameraState createState() =>
+      _WorkoutPortraitStepFinishCameraState();
 }
 
-class _WorkoutLandscapeStepCountingCameraState
-    extends State<WorkoutLandscapeStepCountingCamera> {
+class _WorkoutPortraitStepFinishCameraState
+    extends State<WorkoutPortraitStepFinishCamera> {
   CameraController? _controller;
 
   @override
@@ -43,7 +44,7 @@ class _WorkoutLandscapeStepCountingCameraState
     super.dispose();
   }
 
-  Widget _landscapeMode() {
+  Widget _portraitMode() {
     SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersiveSticky);
     return Column(
       children: [
@@ -51,73 +52,33 @@ class _WorkoutLandscapeStepCountingCameraState
           children: <Widget>[
             Container(
                 decoration: new BoxDecoration(color: color_dark),
-                height: MediaQuery.of(context).size.height * 0.25,
+                height: MediaQuery.of(context).size.height * 0.14,
                 width: MediaQuery.of(context).size.width,
                 child: LinearProgressIndicator(
-                  value: 0.3,
+                  value: 1,
                   valueColor: AlwaysStoppedAnimation(color_dimmedTeal),
                   backgroundColor: color_dark,
                   semanticsLabel: 'Linear progress indicator',
                 )),
             Positioned(
-              top: (MediaQuery.of(context).size.height * 0.25) / 10,
-              right: 60,
-              child: Text("98",
-                  style: const TextStyle(
-                      color: Colors.white,
-                      fontWeight: FontWeight.w600,
-                      fontFamily: "Poppins",
-                      fontStyle: FontStyle.normal,
-                      fontSize: 36.0),
-                  textAlign: TextAlign.center),
-            ),
-            Positioned(
-              bottom: (MediaQuery.of(context).size.height * 0.25) / 10,
-              right: 50,
-              child: Text("99",
-                  style: const TextStyle(
-                      color: color_subtitle,
-                      fontWeight: FontWeight.w400,
-                      fontFamily: "Poppins",
-                      fontStyle: FontStyle.normal,
-                      fontSize: 14.0),
-                  textAlign: TextAlign.center),
-            ),
-            Positioned(
-                top: (MediaQuery.of(context).size.height * 0.25) / 4,
-                left: 25,
-                child: Container(
-                    height: MediaQuery.of(context).size.height / 8,
-                    child: Center(
-                      child: Text("Chest Stretch",
-                          style: const TextStyle(
-                              color: color_dark,
-                              fontWeight: FontWeight.w600,
-                              fontFamily: "Poppins",
-                              fontStyle: FontStyle.normal,
-                              fontSize: 20.0),
-                          textAlign: TextAlign.center),
-                    ))),
-            Positioned(
-              bottom: 0,
-              left: MediaQuery.of(context).size.width * 0.5,
-              child: Center(
-                  child: SvgPicture.asset(
-                'assets/Icon/Detail Expand Icon.svg', // dot dot dot
-                height: 26,
-              )),
-            )
+                top: (MediaQuery.of(context).size.height * 0.14) / 5,
+                left: (MediaQuery.of(context).size.width * 0.4),
+                child: Icon(
+                  Ionicons.checkmark_circle,
+                  size: 70,
+                  color: color_dark,
+                )),
           ],
         ),
         Stack(children: <Widget>[
           Container(
             decoration: new BoxDecoration(color: color_white),
             height: MediaQuery.of(context).size.height -
-                (MediaQuery.of(context).size.height * 0.25),
+                (MediaQuery.of(context).size.height * 0.14),
             width: MediaQuery.of(context).size.width,
           ),
           Positioned(
-              bottom: 15,
+              bottom: 47,
               left: 25,
               child: Stack(children: <Widget>[
                 Container(
@@ -138,7 +99,7 @@ class _WorkoutLandscapeStepCountingCameraState
     );
   }
 
-  Widget _portraitMode() {
+  Widget _landscapeMode() {
     return SafeArea(
         child: new Container(
             decoration: new BoxDecoration(color: color_white),
@@ -146,7 +107,7 @@ class _WorkoutLandscapeStepCountingCameraState
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
                 SvgPicture.asset(
-                  'assets/Icon/Miscellaneous-Outline_phone_hor.svg', // dot dot dot
+                  'assets/Icon/Miscellaneous-Outline_phone.svg', // dot dot dot
                   height: 100,
                 ),
                 Container(
@@ -154,8 +115,8 @@ class _WorkoutLandscapeStepCountingCameraState
                 ),
                 Center(
                     child: Container(
-                  width: MediaQuery.of(context).size.width * 0.6,
-                  child: Text("Please rotate your device horizontally",
+                  width: MediaQuery.of(context).size.width * 0.3,
+                  child: Text("Please rotate your device vertically",
                       style: const TextStyle(
                           color: color_dark,
                           fontWeight: FontWeight.w600,
